+++ app/component/MapComponent.jsx
@@ -1,11 +1,10 @@
-import React, { useCallback, useRef, useState, useEffect } from 'react';
-import { GoogleMap, LoadScript, StandaloneSearchBox, Marker } from '@react-google-maps/api';
-import { Wrapper, Status } from "@googlemaps/react-wrapper";
+import React, { useCallback, useRef, useState } from 'react';
+import { GoogleMap, LoadScript, StandaloneSearchBox, Marker, InfoWindow } from '@react-google-maps/api';
 
 const libraries = ["places"];
 const mapContainerStyle = {
-  height: '400px',
-  width: '100%',
+  height: '600px',
+  width: '70%',
 };
 
 const center = {
@@ -13,85 +12,56 @@ const center = {
   lng: 139.6917, // 東京の経度
 };
 
-const Map = () => {
+const MapWithSearchBox = () => {
   const [map, setMap] = useState(null);
+  const [places, setPlaces] = useState([]);
+  const [selectedPlace, setSelectedPlace] = useState(null);
   const [markers, setMarkers] = useState([]);
   const searchBox = useRef(null);
-  const [customIcon, setCustomIcon] = useState(null);
-  const [isApiLoaded, setIsApiLoaded] = useState(false);
 
-  const onMapLoad = useCallback((mapInstance) => {
-    setMap(mapInstance);
+  const onLoad = useCallback((map) => {
+    setMap(map);
   }, []);
 
   const onPlacesChanged = () => {
-    const places = searchBox.current.getPlaces();
-    if (places.length > 0) {
-      const place = places[0];
+    const newPlaces = searchBox.current.getPlaces();
+    setPlaces(newPlaces);
+    if (newPlaces.length > 0) {
+      const place = newPlaces[0];
       const location = place.geometry.location;
-      const newMarker = { lat: location.lat(), lng: location.lng() };
       map.panTo(location);
-      setMarkers((prevMarkers) => [...prevMarkers, newMarker]);
-    }
-  };
-
-  useEffect(() => {
-    if (isApiLoaded) {
-      setCustomIcon({
-        url: 'https://path-to-your-icon.png', // カスタムアイコンのURL
-        scaledSize: new window.google.maps.Size(50, 50), // アイコンのサイズを調整
-        origin: new window.google.maps.Point(0, 0), // アイコンの起点
-        anchor: new window.google.maps.Point(25, 25), // アイコンのアンカー
-      });
-    }
-  }, [isApiLoaded]);
-
-  const handleApiLoad = () => {
-    setIsApiLoaded(true);
-  };
-
-  const Marker = (options) => {
-    const [marker, setMarker] = React.useState();
-  
-    React.useEffect(() => {
-      if (!marker) {
-        setMarker(new google.maps.Marker());
-      }
-  
-      // remove marker from map on unmount
-      return () => {
-        if (marker) {
-          marker.setMap(null);
-        }
+      const newMarker = {
+        name: place.name,
+        position: {
+          lat: location.lat(),
+          lng: location.lng(),
+        },
       };
-    }, [marker]);
-    React.useEffect(() => {
-      if (marker) {
-        marker.setOptions(options);
-      }
-    }, [marker, options]);
-    return null;
+      setMarkers([...markers, newMarker]);
+      setSelectedPlace(newMarker);
+    }
   };
-  
+
+  const handleMarkerClick = (marker) => {
+    setSelectedPlace(marker);
+  };
+
+  const handleMapClick = () => {
+    setSelectedPlace(null);
+  };
+
   return (
     <LoadScript
-      googleMapsApiKey="AIzaSyBk34MJtHoEDYT7xu9SW4Jl4ctJZP9JZ-U" // ここにAPIキーを入力
+      googleMapsApiKey="AIzaSyBk34MJtHoEDYT7xu9SW4Jl4ctJZP9JZ-U"
       libraries={libraries}
-      onLoad={handleApiLoad}
     >
       <GoogleMap
         mapContainerStyle={mapContainerStyle}
         center={center}
         zoom={10}
-        onLoad={onMapLoad}
+        onLoad={onLoad}
+        onClick={handleMapClick} // マップの空白部分をクリックしたときの処理を追加
       >
-        {markers.map((marker, index) => (
-          <Marker
-            key={index}
-            position={{ lat: marker.lat, lng: marker.lng }}
-            icon={customIcon}
-          />
-        ))}
         <StandaloneSearchBox
           onLoad={(ref) => (searchBox.current = ref)}
           onPlacesChanged={onPlacesChanged}
@@ -109,16 +79,42 @@ const Map = () => {
               boxShadow: "0 2px 6px rgba(0, 0, 0, 0.3)",
               fontSize: "14px",
               outline: "none",
-              textOverflow: "ellipses",
+              textOverflow: "ellipsis",
               position: "absolute",
               left: "50%",
               marginLeft: "-120px",
             }}
           />
         </StandaloneSearchBox>
+        {markers.map((marker, index) => (
+          <Marker
+            key={index}
+            position={marker.position}
+            onClick={() => handleMarkerClick(marker)}
+          >
+            {selectedPlace === marker && (
+              <InfoWindow
+                position={marker.position}
+                onCloseClick={() => setSelectedPlace(null)}
+              >
+                <div>
+                  <h4>{marker.name}</h4>
+                </div>
+              </InfoWindow>
+            )}
+          </Marker>
+        ))}
       </GoogleMap>
     </LoadScript>
   );
 };
 
-export default Map;
+const App = () => {
+  return (
+    <>
+      <MapWithSearchBox />
+    </>
+  );
+};
+
+export default App;
