+++ app/routes/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MapComponent from '../component/MAP';
+import MapComponent from '../component/MapComponent';
 
 function App() {
   return (
